requirements.txt
fastapi==0.85.1
uvicorn==0.27.0
python-dotenv==1.0.0
openai==1.12.0
snowflake-connector-python==3.7.0
google-cloud-secret-manager==2.18.0
pinecone-client==3.0.2
python-multipart==0.0.9
chromadb==0.3.29
pydantic==1.10.13
google-cloud-logging==3.9.0
mistralai==0.0.11
boto3==1.34.34
serpapi==0.1.5





requests
asyncio==3.4.3
tiktoken

boto3
pandas
numpy
spacy


webdriver-manager
selenium
pinecone
webdriver-manager==4.0.1
scikit-learn
streamlit
plotly==6.0.1
tabulate>=0.9.0  
reportlab
pandas
matplotlib
seaborn
scipy
scikit-learn
scikit-fuzzy
kaleido
snowflake-connector-python>=3.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
pandas>=2.2.0
plotly>=5.18.0

langchain>=0.1.0

faiss-cpu>=1.7.4
google-cloud-logging>=3.9.0
google-cloud-storage>=2.11.0
google-cloud-aiplatform>=1.29.0
google-cloud-bigquery>=3.11.0
google-cloud-bigquery-storage>=2.11.0
google-cloud-pubsub>=2.11.0
